+++ spec/hatchet/python_update_warning_spec.rb
@@ -58,18 +58,20 @@ end
 RSpec.describe 'Python update warnings' do
   context 'with a runtime.txt containing python-3.7.12' do
     let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.7_outdated', allow_failure: allow_failure) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.7_outdated', allow_failure:) }
 
-    context 'when using Heroku-18 or Heroku-20', stacks: %w[heroku-18 heroku-20] do
+    context 'when using Heroku-20', stacks: %w[heroku-20] do
       it 'warns about both the deprecated major version and the patch update' do
         app.deploy do |app|
           expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
             remote: -----> Python app detected
             remote: -----> Using Python version specified in runtime.txt
             remote:  !     
-            remote:  !     Python 3.7 will reach its upstream end-of-life on June 27th, 2023, at which
-            remote:  !     point it will no longer receive security updates:
+            remote:  !     Python 3.7 reached its upstream end-of-life on June 27th, 2023, so no longer
+            remote:  !     receives any security updates:
             remote:  !     https://devguide.python.org/versions/#supported-versions
+            remote:  !     
+            remote:  !     Support for Python 3.7 will be removed from this buildpack in October 2023.
             remote:  !     
             remote:  !     Upgrade to a newer Python version as soon as possible to keep your app secure.
             remote:  !     See: https://devcenter.heroku.com/articles/python-runtimes
@@ -93,9 +95,9 @@ RSpec.describe 'Python update warnings' do
 
   context 'with a runtime.txt containing python-3.8.12' do
     let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.8_outdated', allow_failure: allow_failure) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.8_outdated', allow_failure:) }
 
-    context 'when using Heroku-18 or Heroku-20', stacks: %w[heroku-18 heroku-20] do
+    context 'when using Heroku-20', stacks: %w[heroku-20] do
       include_examples 'warns there is a Python update available', '3.8.12', LATEST_PYTHON_3_8
     end
 
@@ -114,14 +116,14 @@ RSpec.describe 'Python update warnings' do
 
   context 'with a runtime.txt containing python-3.10.5' do
     let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.10_outdated', allow_failure: allow_failure) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.10_outdated', allow_failure:) }
 
     include_examples 'warns there is a Python update available', '3.10.5', LATEST_PYTHON_3_10
   end
 
   context 'with a runtime.txt containing python-3.11.0' do
     let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.11_outdated', allow_failure: allow_failure) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.11_outdated', allow_failure:) }
 
     include_examples 'warns there is a Python update available', '3.11.0', LATEST_PYTHON_3_11
   end