+++ spec/spec_helper.rb
@@ -6,16 +6,26 @@ ENV['HATCHET_DEFAULT_STACK'] ||= 'heroku-22'
 require 'rspec/core'
 require 'hatchet'
 
-LATEST_PYTHON_3_7 = '3.7.16'
-LATEST_PYTHON_3_8 = '3.8.16'
-LATEST_PYTHON_3_9 = '3.9.16'
-LATEST_PYTHON_3_10 = '3.10.10'
-LATEST_PYTHON_3_11 = '3.11.2'
+LATEST_PYTHON_3_7 = '3.7.17'
+LATEST_PYTHON_3_8 = '3.8.17'
+LATEST_PYTHON_3_9 = '3.9.17'
+LATEST_PYTHON_3_10 = '3.10.12'
+LATEST_PYTHON_3_11 = '3.11.4'
 DEFAULT_PYTHON_VERSION = LATEST_PYTHON_3_11
 
-PIP_VERSION = '23.0.1'
-SETUPTOOLS_VERSION = '63.4.3'
-WHEEL_VERSION = '0.38.4'
+# The requirement versions are effectively buildpack constants, however, we want
+# Dependabot to be able to update them, which requires that they be in requirements
+# files. The requirements files contain contents like `package==1.2.3` (and not just
+# the package version) so we have to extract the version substring from it.
+def get_requirement_version(package_name)
+  requirement = File.read("requirements/#{package_name}.txt").strip
+  requirement.delete_prefix("#{package_name}==")
+end
+
+PIP_VERSION = get_requirement_version('pip')
+SETUPTOOLS_VERSION = get_requirement_version('setuptools')
+WHEEL_VERSION = get_requirement_version('wheel')
+PIPENV_VERSION = get_requirement_version('pipenv')
 
 # Work around the return value for `default_buildpack` changing after deploy:
 # https://github.com/heroku/hatchet/issues/180